requirements.txt
mysql.connector
faker